rev(specman-core): workspace discovery
This aims to provide requirements for implementers to give the ability to find relationships from file paths, as well as resolve local workspaces.
Along with this as a sort of experimental method, I added URI protocols spec:// and impl://. After doing it, I questioned the wisdom of it, so this isn't guaranteed to stay. I need to take a look after a day or two and I've had a step back to make sure that this is a good idea. Neither do I know the implications of defining such a protocol, so it does require more thought.