chore(specman-core): relocate constraints
The constraints section provided an area that wasn't necessarily useful as a lot of the constraints are listed inside of the entities or concepts. This aims to move those constraints into corresponding entities and concepts as well as removing the constraints section. This was done with a quick once-over, so there might be some areas missing, though from what I could tell with the checks that I made, I don't see anything out of the ordinary.