feat: version 1.0.0
This marks the general release of Spec-Man, but that doesn't mean that it's finished. Outlined is a roadmap with just the two items that I can think of at the top of my head I want to take care of first. There are a plethora of other quality-of-life things I plan on adding, but since they're something that you sort of think of when you come across them, they're not added to the roadmap.